requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
Pillow>=9.5.0
scikit-image>=0.20.0
albumentations>=1.3.0
lpips>=0.1.4
scipy>=1.10.0
requests>=2.28.0
beautifulsoup4>=4.12.0
icrawler>=0.6.6
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
PyYAML>=6.0
matplotlib>=3.7.0
scikit-learn>=1.2.0